requirements.txt
numpy>=1.20.0
torch>=1.8.0
torchvision>=0.9.0
Pillow>=8.0.0
opencv-python>=4.5.0
albumentations>=1.0.0
matplotlib>=3.3.0
tqdm>=4.50.0
gdown>=4.4.0
tifffile>=2021.0.0
scipy>=1.6.0
pyyaml>=5.4.0
scikit-learn>=0.24.0
requests>=2.25.0